Moving Random h and cpp into the global Common folder
