--- NeuralNetwork/Common/Random.h
@@ -1,8 +1,0 @@
-#ifndef AIENGINE_RANDOM_H
-#define AIENGINE_RANDOM_H
-namespace random
-{
-    void initRandomSeed();
-    float range(float a, float b);
-}
-#endif //AIENGINE_RANDOM_H
