+++ NeuralNetwork/MLP/CMakeLists.txt
@@ -1,7 +1,7 @@
 set(FILES
         ../Common/ActivationFunctions.cpp
-        ../Common/Random.cpp
         ../Common/Matrix.cpp
+        ../../Common/Random.cpp
         ../Perceptron/Neuron.cpp
         ../Perceptron/Perceptron.cpp
         Layer.cpp
