+++ NeuralNetwork/Perceptron/CMakeLists.txt
@@ -1,7 +1,7 @@
 set(FILES
         ../Common/ActivationFunctions.cpp
-        ../Common/Random.cpp
         ../Common/Matrix.cpp
+        ../../Common/Random.cpp
         Neuron.cpp
         Perceptron.cpp
         PerceptronTests.cpp
